split crop test module + ruff fixes & formats
