--- backend/tests/scrapers/test_scraper_crops.py
@@ -1,1813 +1,0 @@
-"""
-Tests for the media-item crop section of the Viernulvier scraper.
-
-Covers every line of:
-  - _derive_crop_filename
-  - _download_image
-  - sync_media_item_crops
-
-Run with:
-    pytest apps/imports/scrapers/tests/test_viernulvier_crops.py -v
-"""
-
-from __future__ import annotations
-
-from datetime import UTC, datetime
-import logging
-from unittest.mock import Mock
-
-from django.core.exceptions import FieldDoesNotExist
-import pytest
-import requests
-
-from apps.import_log.models import ImportLog
-from apps.imports.scrapers import viernulvier
-from apps.imports.scrapers.viernulvier import (
-    ScraperError,
-    _derive_crop_filename,
-    _download_image,
-    sync_media_item_crops,
-)
-from apps.media_library import models as media_models
-
-# ---------------------------------------------------------------------------
-# Helpers shared across tests
-# ---------------------------------------------------------------------------
-
-
-def _make_ok_response(content: bytes = b"img-bytes", status: int = 200):
-    """Return a mock requests.Response with .ok=True and .content set."""
-    r = Mock()
-    r.status_code = status
-    r.ok = True
-    r.content = content
-    r.headers = Mock()
-    r.headers.get = Mock(return_value=None)
-    return r
-
-
-def _make_status_response(status: int, headers: dict | None = None):
-    r = Mock()
-    r.status_code = status
-    r.ok = status < 400
-    r.headers = Mock()
-    header_dict = headers or {}
-    r.headers.get = lambda k, d=None: header_dict.get(k, d)
-    return r
-
-
-# ============================================================================
-# _derive_crop_filename
-# ============================================================================
-
-
-class TestDeriveCropFilename:
-    def test_basic_path_and_jpg_extension(self) -> None:
-        """/api/v1/media/items/310 + .jpg url -> api_v1_media_items_310_hd_ready.jpg"""
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/310",
-            "https://cdn.example.com/images/photo.jpg",
-        )
-        assert result == "api_v1_media_items_310_hd_ready.jpg"
-
-    def test_extension_extracted_from_png_url(self) -> None:
-        result = _derive_crop_filename(
-            "FE3_header",
-            "/api/v1/media/items/42",
-            "https://cdn.example.com/images/banner.png",
-        )
-        assert result.endswith(".png")
-        assert "FE3_header" in result
-
-    def test_query_string_stripped_before_extension_extraction(self) -> None:
-        """Query params after ? are ignored when detecting the extension."""
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/5",
-            "https://cdn.example.com/photo.jpeg?w=800&h=600",
-        )
-        assert result.endswith(".jpeg")
-
-    def test_extension_too_long_falls_back_to_jpg(self) -> None:
-        """Extension with more than 5 chars (e.g. a CDN hash) falls back to .jpg."""
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/9",
-            "https://cdn.example.com/images/abc123def456",  # no dot -> falls back
-        )
-        assert result.endswith(".jpg")
-
-    def test_non_alpha_extension_falls_back_to_jpg(self) -> None:
-        """Extension containing digits (like .123) falls back to .jpg."""
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/7",
-            "https://cdn.example.com/file.12345",
-        )
-        assert result.endswith(".jpg")
-
-    def test_url_segment_without_dot_falls_back_to_jpg(self) -> None:
-        """A URL path that has no dot at all falls back to .jpg."""
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/8",
-            "https://cdn.example.com/images/nondotted",
-        )
-        assert result.endswith(".jpg")
-
-    def test_leading_slash_stripped_from_external_id(self) -> None:
-        """Leading slash on external_id is stripped before slugifying."""
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/10",
-            "https://cdn.example.com/img.jpg",
-        )
-        # Must NOT start with underscore
-        assert not result.startswith("_")
-        assert result.startswith("api")
-
-    def test_slashes_in_external_id_replaced_by_underscores(self) -> None:
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/310",
-            "https://cdn.example.com/img.jpg",
-        )
-        assert "/" not in result
-
-    def test_crop_name_included_in_filename(self) -> None:
-        result = _derive_crop_filename(
-            "FE3_header",
-            "/api/v1/media/items/99",
-            "https://cdn.example.com/img.jpg",
-        )
-        assert "FE3_header" in result
-
-    def test_extension_longer_than_5_chars_falls_back(self) -> None:
-        """Any extension longer than 5 chars (including the dot) falls back to .jpg."""
-        # ".webp2" is 6 chars -> fallback
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/1",
-            "https://cdn.example.com/img.webp2",
-        )
-        assert result.endswith(".jpg")
-
-    def test_webp_extension_is_kept(self) -> None:
-        """.webp is exactly 5 chars - border case that should be kept."""
-        result = _derive_crop_filename(
-            "hd_ready",
-            "/api/v1/media/items/1",
-            "https://cdn.example.com/img.webp",
-        )
-        assert result.endswith(".webp")
-
-
-# ============================================================================
-# _download_image
-# ============================================================================
-
-
-class TestDownloadImage:
-    def _patched_session(self, monkeypatch, responses_fn):
-        """Return a Mock session whose .get() calls responses_fn(url, call_n)."""
-        call_count = [0]
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-
-        def fake_get(url, timeout=None, stream=None):
-            call_count[0] += 1
-            return responses_fn(url, call_count[0])
-
-        session.get.side_effect = fake_get
-        return session, call_count
-
-    def test_returns_bytes_on_success(self, monkeypatch) -> None:
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-        session.get.return_value = _make_ok_response(b"pixels")
-
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result == b"pixels"
-
-    def test_returns_none_on_non_ok_status(self, monkeypatch) -> None:
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-        session.get.return_value = _make_status_response(404)
-
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result is None
-
-    def test_returns_none_on_request_exception(self, monkeypatch) -> None:
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-        session.get.side_effect = requests.RequestException("ssl error")
-
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result is None
-
-    def test_connection_error_retries_then_succeeds(self, monkeypatch) -> None:
-        def responses(url, n):
-            if n == 1:
-                raise requests.ConnectionError("down")
-            return _make_ok_response(b"data")
-
-        session, count = self._patched_session(monkeypatch, responses)
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result == b"data"
-        assert count[0] == 2
-
-    def test_timeout_retries_then_succeeds(self, monkeypatch) -> None:
-        def responses(url, n):
-            if n == 1:
-                raise requests.Timeout
-            return _make_ok_response(b"ok")
-
-        session, _ = self._patched_session(monkeypatch, responses)
-        assert _download_image(session, "https://cdn.example.com/img.jpg") == b"ok"
-
-    def test_connection_error_all_retries_exhausted_returns_none(self, monkeypatch, caplog) -> None:
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-        session.get.side_effect = requests.ConnectionError("always down")
-
-        caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result is None
-        assert any("Image download failed" in r.message for r in caplog.records)
-
-    def test_timeout_all_retries_exhausted_returns_none(self, monkeypatch, caplog) -> None:
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-        session.get.side_effect = requests.Timeout()
-
-        caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result is None
-
-    def test_429_with_retry_after_waits_then_succeeds(self, monkeypatch) -> None:
-        sleep_calls = []
-        monkeypatch.setattr(viernulvier.time, "sleep", sleep_calls.append)
-        call_count = [0]
-        session = Mock()
-
-        def fake_get(url, timeout=None, stream=None):
-            call_count[0] += 1
-            if call_count[0] == 1:
-                return _make_status_response(429, {"Retry-After": "10"})
-            return _make_ok_response(b"img")
-
-        session.get.side_effect = fake_get
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result == b"img"
-        assert 10.0 in sleep_calls
-
-    def test_429_all_retries_exhausted_returns_none(self, monkeypatch, caplog) -> None:
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-        session.get.return_value = _make_status_response(429, {})
-
-        caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result is None
-        assert any("Rate limited" in r.message for r in caplog.records)
-
-    def test_429_without_retry_after_uses_backoff(self, monkeypatch) -> None:
-        sleep_calls = []
-        monkeypatch.setattr(viernulvier.time, "sleep", sleep_calls.append)
-        call_count = [0]
-        session = Mock()
-
-        def fake_get(url, timeout=None, stream=None):
-            call_count[0] += 1
-            if call_count[0] == 1:
-                return _make_status_response(429, {})
-            return _make_ok_response(b"ok")
-
-        session.get.side_effect = fake_get
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result == b"ok"
-        # Some backoff sleep must have been called (not the fixed 10s from Retry-After)
-        assert sleep_calls
-
-    def test_server_error_non_ok_non_429_logs_and_returns_none(self, monkeypatch, caplog) -> None:
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-        session.get.return_value = _make_status_response(500)
-
-        caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result is None
-        assert any("HTTP 500" in r.message for r in caplog.records)
-
-    def test_fallthrough_return_none_when_max_retries_set_to_minus_one(self, monkeypatch) -> None:
-        """The unreachable `return None` after the retry loop is hit when MAX_RETRIES=-1.
-
-        With MAX_RETRIES=-1 the for-loop body never executes, so the function
-        falls through to the bare `return None` at the bottom.
-        """
-        monkeypatch.setattr(viernulvier, "MAX_RETRIES", -1)
-        session = Mock()  # .get should never be called
-
-        result = _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert result is None
-        session.get.assert_not_called()
-
-    def test_connection_error_retry_warning_logged(self, monkeypatch, caplog) -> None:
-        """A connection error that succeeds on retry logs a warning."""
-        call_count = [0]
-        monkeypatch.setattr(viernulvier.time, "sleep", lambda *_: None)
-        session = Mock()
-
-        def fake_get(url, timeout=None, stream=None):
-            call_count[0] += 1
-            if call_count[0] == 1:
-                raise requests.ConnectionError("flaky")
-            return _make_ok_response(b"data")
-
-        session.get.side_effect = fake_get
-        caplog.set_level(logging.WARNING, logger=viernulvier.logger.name)
-        _download_image(session, "https://cdn.example.com/img.jpg")
-
-        assert any("Image download error" in r.message for r in caplog.records)
-
-
-# ============================================================================
-# sync_media_item_crops
-# ============================================================================
-
-
-def _patch_crop_dependencies(
-    monkeypatch,
-    *,
-    foto_items: list,
-    fetch_results: dict | None = None,
-    download_result: bytes | None = b"image-data",
-    synced_crop_names: set | None = None,
-    generate_filename_result: str = "media/crops/item_hd_ready.jpg",
-    storage_save_result: str = "media/crops/item_hd_ready.jpg",
-):
-    """
-    Patch all external dependencies of sync_media_item_crops in one call.
-
-    fetch_results: maps external_id -> (item_data, etag) tuple.
-    download_result: bytes returned by _download_image (None simulates failure).
-    """
-    if synced_crop_names is None:
-        synced_crop_names = {"hd_ready", "FE3_header"}
-
-    # --- MediaItem queryset ---
-    mock_qs = Mock()
-    mock_qs.values.return_value = foto_items
-    monkeypatch.setattr(
-        media_models.MediaItem.objects,
-        "filter",
-        lambda **_kw: mock_qs,
-    )
-
-    # --- _build_session (returns a dummy session; _download_image is patched separately) ---
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-
-    # --- _fetch_with_retry ---
-    def fake_fetch(session, url, etag=None, params=None):
-        if fetch_results and url in fetch_results:
-            return fetch_results[url]
-        # Default: single crop
-        return (
-            {
-                "crops": [
-                    {"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"},
-                ]
-            },
-            None,
-        )
-
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", fake_fetch)
-
-    # --- _download_image ---
-    monkeypatch.setattr(viernulvier, "_download_image", lambda _session, _url: download_result)
-
-    # --- MediaItemCrop ---
-    mock_crop_model = Mock()
-    mock_crop_model.SYNCED_CROP_NAMES = synced_crop_names
-
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = generate_filename_result
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = storage_save_result
-    mock_crop_model._meta = Mock()
-    mock_crop_model._meta.get_field.return_value = mock_image_field
-    mock_crop_model.objects.update_or_create.return_value = (Mock(), True)
-
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_model)
-
-    return mock_crop_model
-
-
-@pytest.mark.django_db
-def test_sync_crops_no_foto_items_returns_zero(monkeypatch) -> None:
-    """When there are no foto MediaItems, sync returns 0 immediately."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = []
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-
-    result = sync_media_item_crops()
-
-    assert result == 0
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skip_missing_fields_and_apply_supported_ones(monkeypatch) -> None:
-    class FakeQuerySet:
-        def __init__(self) -> None:
-            self.filter_calls = []
-
-        def filter(self, **kwargs):
-            self.filter_calls.append(kwargs)
-            return self
-
-        def values(self, *_args, **_kwargs):
-            return []
-
-    fake_qs = FakeQuerySet()
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: fake_qs)
-
-    def fake_get_field(name):
-        if name == "updated_at":
-            return Mock()
-        raise FieldDoesNotExist(name)
-
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", fake_get_field)
-
-    def fake_parse_datetime(value):
-        if value == "raise-type":
-            raise TypeError("bad type")
-        return datetime(2024, 1, 1, tzinfo=UTC)
-
-    monkeypatch.setattr(viernulvier, "parse_datetime", fake_parse_datetime)
-    monkeypatch.setattr(viernulvier, "fetch_viernulvier", lambda **_kw: [])
-
-    result = sync_media_item_crops(
-        params={
-            "created_at[after]": "ok",
-            "created_at[before]": "ok",
-            "updated_at[after]": "ok",
-            "updated_at[before]": "ok",
-            "updated_at[strictly_after]": "raise-type",
-            "updated_at[strictly_before]": "ok",
-            "created_at[strictly_after]": "ok",
-            "created_at[strictly_before]": "raise-value",
-        }
-    )
-
-    assert result == 0
-    assert {"updated_at__gte": datetime(2024, 1, 1, tzinfo=UTC)} in fake_qs.filter_calls
-    assert {"updated_at__lte": datetime(2024, 1, 1, tzinfo=UTC)} in fake_qs.filter_calls
-    assert all("created_at" not in next(iter(call)) for call in fake_qs.filter_calls)
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_are_forwarded_to_api_and_limit_item_fetches(monkeypatch) -> None:
-    class FakeQuerySet:
-        def values(self, *_args, **_kwargs):
-            return [
-                {"pk": 1, "external_id": "/api/v1/media/items/1"},
-                {"pk": 2, "external_id": "/api/v1/media/items/2"},
-            ]
-
-    fake_qs = FakeQuerySet()
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: fake_qs)
-
-    def fake_get_field(name):
-        if name == "updated_at":
-            return Mock()
-        raise FieldDoesNotExist(name)
-
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", fake_get_field)
-    monkeypatch.setattr(viernulvier, "parse_datetime", Mock(return_value=None))
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-
-    captured = {}
-
-    def fake_fetch_viernulvier(endpoint, params=None, etag_cache=None):
-        captured["endpoint"] = endpoint
-        captured["params"] = params
-        return [{"@id": "/api/v1/media/items/1"}]
-
-    monkeypatch.setattr(viernulvier, "fetch_viernulvier", fake_fetch_viernulvier)
-
-    fetched_urls = []
-
-    def fake_fetch_with_retry(session, url, params=None, etag=None):
-        fetched_urls.append(url)
-        return ({"crops": []}, None)
-
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", fake_fetch_with_retry)
-
-    result = sync_media_item_crops(
-        params={
-            "created_at[after]": "2024-01-01T00:00:00+00:00",
-            "updated_at[after]": "2024-01-01T00:00:00+00:00",
-        }
-    )
-
-    assert result == 0
-    assert captured["endpoint"] == "/media/items"
-    assert captured["params"] == {
-        "created_at[after]": "2024-01-01T00:00:00+00:00",
-        "updated_at[after]": "2024-01-01T00:00:00+00:00",
-    }
-    assert fetched_urls == ["https://www.viernulvier.gent/api/v1/media/items/1"]
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skip_non_image_api_items(monkeypatch) -> None:
-    class FakeQuerySet:
-        def values(self, *_args, **_kwargs):
-            return []
-
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: FakeQuerySet())
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", lambda name: Mock() if name == "updated_at" else None)
-    monkeypatch.setattr(viernulvier, "parse_datetime", Mock(return_value=None))
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [
-            {"@id": "/api/v1/media/items/vid-1", "type": "video"},
-            {"@id": "/api/v1/media/items/img-1", "type": "foto"},
-        ],
-    )
-
-    fetched_urls = []
-
-    def fake_fetch_with_retry(session, url, params=None, etag=None):
-        fetched_urls.append(url)
-        return ({"crops": []}, None)
-
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", fake_fetch_with_retry)
-
-    result = sync_media_item_crops(params={"updated_at[after]": "2024-01-01T00:00:00+00:00"})
-
-    assert result == 0
-    assert fetched_urls == ["https://www.viernulvier.gent/api/v1/media/items/img-1"]
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skip_api_items_without_external_id(monkeypatch) -> None:
-    class FakeQuerySet:
-        def filter(self, **_kwargs):
-            return self
-
-        def values(self, *_args, **_kwargs):
-            return [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: FakeQuerySet())
-
-    def fake_get_field(_name):
-        return Mock()
-
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", fake_get_field)
-    monkeypatch.setattr(viernulvier, "parse_datetime", Mock(return_value=datetime(2024, 1, 1, tzinfo=UTC)))
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [{"@id": "", "type": "foto"}],
-    )
-
-    result = sync_media_item_crops(params={"updated_at[after]": "2024-01-01T00:00:00+00:00"})
-
-    assert result == 0
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skip_non_dict_api_items(monkeypatch) -> None:
-    class FakeQuerySet:
-        def values(self, *_args, **_kwargs):
-            return []
-
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: FakeQuerySet())
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", lambda name: Mock() if name == "updated_at" else None)
-    monkeypatch.setattr(viernulvier, "parse_datetime", Mock(return_value=None))
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [
-            "not-a-dict",
-            {"@id": "/api/v1/media/items/img-2", "type": "foto"},
-        ],
-    )
-
-    fetched_urls = []
-
-    def fake_fetch_with_retry(session, url, params=None, etag=None):
-        fetched_urls.append(url)
-        return ({"crops": []}, None)
-
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", fake_fetch_with_retry)
-
-    result = sync_media_item_crops(params={"updated_at[after]": "2024-01-01T00:00:00+00:00"})
-
-    assert result == 0
-    assert fetched_urls == ["https://www.viernulvier.gent/api/v1/media/items/img-2"]
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_fetch_error_marks_import_log_failed(monkeypatch) -> None:
-    class FakeQuerySet:
-        def values(self, *_args, **_kwargs):
-            return []
-
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: FakeQuerySet())
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", lambda name: Mock() if name == "updated_at" else None)
-    monkeypatch.setattr(viernulvier, "parse_datetime", Mock(return_value=None))
-
-    monkeypatch.setattr(viernulvier, "fetch_viernulvier", Mock(side_effect=RuntimeError("api down")))
-
-    with pytest.raises(RuntimeError, match="api down"):
-        sync_media_item_crops(params={"updated_at[after]": "2024-01-01T00:00:00+00:00"})
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.status == ImportLog.Status.FAILED
-    assert log.error_message == "api down"
-    assert log.finished_at is not None
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_upserts_missing_media_item_dependency(monkeypatch) -> None:
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [{"@id": "/api/v1/media/items/42", "type": "foto"}],
-    )
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (
-            {
-                "type": "foto",
-                "original_filename": "missing.jpg",
-                "position": 0,
-                "format": "jpg",
-                "crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}],
-            },
-            None,
-        ),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"img")
-
-    media_upsert = Mock(return_value=(Mock(pk=4242), True))
-    monkeypatch.setattr(media_models.MediaItem.objects, "update_or_create", media_upsert)
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "crops/img.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.return_value = (Mock(), True)
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    result = sync_media_item_crops(params={"updated_at[after]": "2024-01-01T00:00:00+00:00"})
-
-    assert result == 1
-    media_upsert.assert_called_once()
-    mock_crop_cls.objects.update_or_create.assert_called_once_with(
-        media_item_id=4242,
-        name="hd_ready",
-        defaults={"image": "crops/img.jpg"},
-    )
-
-
-@pytest.mark.django_db
-def test_sync_crops_upsert_missing_media_item_resolves_gallery_id(monkeypatch) -> None:
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [{"@id": "/api/v1/media/items/55", "type": "foto"}],
-    )
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (
-            {
-                "type": "foto",
-                "gallery": "  /api/v1/media/galleries/77  ",
-                "original_filename": "missing.jpg",
-                "position": 0,
-                "format": "jpg",
-                "crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}],
-            },
-            None,
-        ),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"img")
-
-    gallery_qs = Mock()
-    gallery_qs.values_list.return_value.first.return_value = 777
-    gallery_filter = Mock(return_value=gallery_qs)
-    monkeypatch.setattr(media_models.MediaGallery.objects, "filter", gallery_filter)
-
-    media_upsert = Mock(return_value=(Mock(pk=4242), True))
-    monkeypatch.setattr(media_models.MediaItem.objects, "update_or_create", media_upsert)
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "crops/img.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.return_value = (Mock(), True)
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    result = sync_media_item_crops(params={"updated_at[after]": "2024-01-01T00:00:00+00:00"})
-
-    assert result == 1
-    gallery_filter.assert_called_once_with(external_id="/api/v1/media/galleries/77")
-    gallery_qs.values_list.assert_called_once_with("pk", flat=True)
-    assert media_upsert.call_args.kwargs["defaults"]["gallery_id"] == 777
-
-
-@pytest.mark.django_db
-def test_sync_crops_dry_run_logs_missing_media_item_upsert(monkeypatch, caplog) -> None:
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [{"@id": "/api/v1/media/items/123", "type": "foto"}],
-    )
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (
-            {
-                "type": "foto",
-                "crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}],
-            },
-            None,
-        ),
-    )
-
-    media_upsert = Mock()
-    monkeypatch.setattr(media_models.MediaItem.objects, "update_or_create", media_upsert)
-
-    caplog.set_level(logging.INFO, logger=viernulvier.logger.name)
-    result = sync_media_item_crops(
-        dry_run=True,
-        params={"updated_at[after]": "2024-01-01T00:00:00+00:00"},
-    )
-
-    assert result == 0
-    media_upsert.assert_not_called()
-    assert any(
-        "[DRY RUN] Would upsert missing MediaItem for crop sync: /api/v1/media/items/123" in r.message
-        for r in caplog.records
-    )
-
-
-@pytest.mark.django_db
-def test_sync_crops_unresolved_dependency_calls_on_progress(monkeypatch) -> None:
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [{"@id": "/api/v1/media/items/124", "type": "foto"}],
-    )
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"type": "foto", "crops": []}, None),
-    )
-
-    calls = []
-    result = sync_media_item_crops(
-        dry_run=True,
-        params={"updated_at[after]": "2024-01-01T00:00:00+00:00"},
-        on_progress=lambda idx, total: calls.append((idx, total)),
-    )
-
-    assert result == 0
-    assert calls == [(1, 1)]
-
-
-@pytest.mark.django_db
-def test_sync_crops_unresolved_media_item_dependency_after_upsert(monkeypatch, caplog) -> None:
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [{"@id": "/api/v1/media/items/999", "type": "foto"}],
-    )
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (
-            {
-                "type": "foto",
-                "original_filename": "missing.jpg",
-                "position": 0,
-                "format": "jpg",
-                "crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}],
-            },
-            None,
-        ),
-    )
-
-    media_upsert = Mock(return_value=(Mock(pk=None), True))
-    monkeypatch.setattr(media_models.MediaItem.objects, "update_or_create", media_upsert)
-
-    download_image = Mock(return_value=b"img")
-    monkeypatch.setattr(viernulvier, "_download_image", download_image)
-
-    caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-    result = sync_media_item_crops(params={"updated_at[after]": "2024-01-01T00:00:00+00:00"})
-
-    assert result == 0
-    media_upsert.assert_called_once()
-    download_image.assert_not_called()
-    assert any("MediaItem dependency unresolved for crop sync: /api/v1/media/items/999" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_missing_media_item_upsert_exception_is_logged_and_recorded(monkeypatch, caplog) -> None:
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "fetch_viernulvier",
-        lambda **_kw: [{"@id": "/api/v1/media/items/321", "type": "foto"}],
-    )
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (
-            {
-                "type": "foto",
-                "original_filename": "broken.jpg",
-                "position": 0,
-                "format": "jpg",
-                "crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}],
-            },
-            None,
-        ),
-    )
-
-    media_upsert = Mock(side_effect=RuntimeError("upsert exploded"))
-    monkeypatch.setattr(media_models.MediaItem.objects, "update_or_create", media_upsert)
-
-    on_progress = Mock()
-    download_image = Mock(return_value=b"img")
-    monkeypatch.setattr(viernulvier, "_download_image", download_image)
-
-    caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-    result = sync_media_item_crops(
-        params={"updated_at[after]": "2024-01-01T00:00:00+00:00"},
-        on_progress=on_progress,
-    )
-
-    assert result == 0
-    media_upsert.assert_called_once()
-    download_image.assert_not_called()
-    on_progress.assert_called_once_with(1, 1)
-    assert any(
-        "Failed to upsert MediaItem dependency for crop sync (/api/v1/media/items/321)" in r.message for r in caplog.records
-    )
-
-    log = ImportLog.objects.latest("started_at")
-    assert "Missing MediaItem upsert failed for /api/v1/media/items/321" in (log.error_message or "")
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skip_filter_when_fake_parse_datetime_raises_valueerror(monkeypatch) -> None:
-    class FakeQuerySet:
-        def __init__(self) -> None:
-            self.filter_calls = []
-
-        def values(self, *_args, **_kwargs):
-            return []
-
-    fake_qs = FakeQuerySet()
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: fake_qs)
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", lambda name: Mock() if name == "updated_at" else None)
-
-    def fake_parse_datetime(value) -> None:
-        if value == "raise-value":
-            raise ValueError("bad datetime")
-
-    monkeypatch.setattr(viernulvier, "parse_datetime", fake_parse_datetime)
-    monkeypatch.setattr(viernulvier, "fetch_viernulvier", lambda **_kw: [])
-
-    result = sync_media_item_crops(params={"updated_at[strictly_before]": "raise-value"})
-
-    assert result == 0
-    assert fake_qs.filter_calls == []
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skip_filter_when_fake_get_field_raises_fielddoesnotexist(monkeypatch) -> None:
-    class FakeQuerySet:
-        def __init__(self) -> None:
-            self.filter_calls = []
-
-        def filter(self, **kwargs):
-            self.filter_calls.append(kwargs)
-            return self
-
-        def values(self, *_args, **_kwargs):
-            return []
-
-    fake_qs = FakeQuerySet()
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: fake_qs)
-
-    def fake_get_field(name):
-        if name == "updated_at":
-            return Mock()
-        raise FieldDoesNotExist(name)
-
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", fake_get_field)
-
-    parse_dt = Mock(return_value=datetime(2024, 1, 1, tzinfo=UTC))
-    monkeypatch.setattr(viernulvier, "parse_datetime", parse_dt)
-    monkeypatch.setattr(viernulvier, "fetch_viernulvier", lambda **_kw: [])
-
-    result = sync_media_item_crops(
-        params={
-            "created_at[before]": "2024-01-01T00:00:00+00:00",
-            "updated_at[before]": "2024-01-01T00:00:00+00:00",
-        }
-    )
-
-    assert result == 0
-    assert fake_qs.filter_calls == [{"updated_at__lte": datetime(2024, 1, 1, tzinfo=UTC)}]
-    parse_dt.assert_called_once_with("2024-01-01T00:00:00+00:00")
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_ignores_non_matching_param_key(monkeypatch) -> None:
-    class FakeQuerySet:
-        def __init__(self) -> None:
-            self.filter_calls = []
-
-        def values(self, *_args, **_kwargs):
-            return []
-
-    fake_qs = FakeQuerySet()
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: fake_qs)
-
-    meta_get_field = Mock(side_effect=AssertionError("get_field should not be called for invalid param keys"))
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", meta_get_field)
-
-    parse_dt = Mock(side_effect=AssertionError("parse_datetime should not be called for invalid param keys"))
-    monkeypatch.setattr(viernulvier, "parse_datetime", parse_dt)
-
-    result = sync_media_item_crops(params={"invalid[param]": "2024-01-01T00:00:00+00:00"})
-
-    assert result == 0
-    assert fake_qs.filter_calls == []
-    meta_get_field.assert_not_called()
-    parse_dt.assert_not_called()
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skips_filter_when_datetime_is_none(monkeypatch) -> None:
-    class FakeQuerySet:
-        def __init__(self) -> None:
-            self.filter_calls = []
-
-        def values(self, *_args, **_kwargs):
-            return []
-
-    fake_qs = FakeQuerySet()
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: fake_qs)
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", lambda name: Mock() if name == "updated_at" else None)
-
-    parse_dt = Mock(return_value=None)
-    monkeypatch.setattr(viernulvier, "parse_datetime", parse_dt)
-    monkeypatch.setattr(viernulvier, "fetch_viernulvier", lambda **_kw: [])
-
-    result = sync_media_item_crops(params={"updated_at[after]": "not-a-datetime"})
-
-    assert result == 0
-    assert fake_qs.filter_calls == []
-    parse_dt.assert_called_once_with("not-a-datetime")
-
-
-@pytest.mark.django_db
-def test_sync_crops_params_skips_filter_when_datetime_raises_valueerror(monkeypatch) -> None:
-    class FakeQuerySet:
-        def __init__(self) -> None:
-            self.filter_calls = []
-
-        def values(self, *_args, **_kwargs):
-            return []
-
-    fake_qs = FakeQuerySet()
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_kw: fake_qs)
-
-    def fake_get_field(_name):
-        return Mock()
-
-    monkeypatch.setattr(media_models.MediaItem._meta, "get_field", fake_get_field)
-
-    parse_dt = Mock(side_effect=ValueError("bad datetime"))
-    monkeypatch.setattr(viernulvier, "parse_datetime", parse_dt)
-    monkeypatch.setattr(viernulvier, "fetch_viernulvier", lambda **_kw: [])
-
-    result = sync_media_item_crops(params={"updated_at[before]": "raise-value"})
-
-    assert result == 0
-    assert fake_qs.filter_calls == []
-    parse_dt.assert_called_once_with("raise-value")
-
-
-@pytest.mark.django_db
-def test_sync_crops_no_foto_items_creates_success_import_log(monkeypatch) -> None:
-    """Empty foto list creates a SUCCESS ImportLog with all-zero counters."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = []
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.status == ImportLog.Status.SUCCESS
-    assert log.records_total == 0
-    assert log.records_imported == 0
-    assert log.records_failed == 0
-
-
-@pytest.mark.django_db
-def test_sync_crops_item_without_external_id_skipped(monkeypatch, caplog) -> None:
-    """MediaItem rows with no external_id are skipped and counted as errors."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": ""}],
-    )
-
-    caplog.set_level(logging.WARNING, logger=viernulvier.logger.name)
-    result = sync_media_item_crops()
-
-    assert result == 0
-    assert any("no external_id" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_none_external_id_skipped(monkeypatch, caplog) -> None:
-    """MediaItem rows with external_id=None are also skipped."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": None}],
-    )
-
-    caplog.set_level(logging.WARNING, logger=viernulvier.logger.name)
-    result = sync_media_item_crops()
-
-    assert result == 0
-    assert any("no external_id" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_fetch_failure_logged_as_error(monkeypatch, caplog) -> None:
-    """A ScraperError during individual item fetch is logged and counted as error."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (_ for _ in ()).throw(ScraperError("API down")),
-    )
-
-    caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-    result = sync_media_item_crops()
-
-    assert result == 0
-    assert any("Failed to fetch media item" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_304_not_modified_skips_item(monkeypatch) -> None:
-    """A 304 (fetch returns None) means no crops to process - item is silently skipped."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", lambda *_a, **_kw: (None, None))
-
-    result = sync_media_item_crops()
-
-    assert result == 0
-
-
-@pytest.mark.django_db
-def test_sync_crops_non_list_crops_field_skipped(monkeypatch, caplog) -> None:
-    """When API crops field is not a list, the item is silently skipped."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": "not-a-list"}, None),
-    )
-
-    caplog.set_level(logging.DEBUG, logger=viernulvier.logger.name)
-    result = sync_media_item_crops()
-
-    assert result == 0
-    assert any("Unexpected crops format" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_crop_not_in_wanted_set_skipped(monkeypatch) -> None:
-    """Crop names not in SYNCED_CROP_NAMES are silently ignored."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": [{"name": "unknown_variant", "url": "https://cdn.example.com/img.jpg"}]}, None),
-    )
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_qs2 = Mock()
-    mock_qs2.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    result = sync_media_item_crops()
-
-    assert result == 0
-    mock_crop_cls.objects.update_or_create.assert_not_called()
-
-
-@pytest.mark.django_db
-def test_sync_crops_crop_missing_url_warns_and_skips(monkeypatch, caplog) -> None:
-    """A crop dict without a url key is skipped with a warning."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": [{"name": "hd_ready", "url": ""}]}, None),
-    )
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    caplog.set_level(logging.WARNING, logger=viernulvier.logger.name)
-    result = sync_media_item_crops()
-
-    assert result == 0
-    assert any("has no URL" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_download_failure_counted_as_error(monkeypatch, caplog) -> None:
-    """When _download_image returns None the crop is counted as an error."""
-    mock_crop_cls = _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-        download_result=None,  # simulate download failure
-    )
-
-    caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-    result = sync_media_item_crops()
-
-    assert result == 0
-    mock_crop_cls.objects.update_or_create.assert_not_called()
-
-
-@pytest.mark.django_db
-def test_sync_crops_successful_save_returns_count(monkeypatch) -> None:
-    """A successful crop save increments the saved counter."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-        download_result=b"image-bytes",
-    )
-
-    result = sync_media_item_crops()
-
-    assert result == 1
-
-
-@pytest.mark.django_db
-def test_sync_crops_creates_import_log_success(monkeypatch) -> None:
-    """A fully successful run produces a SUCCESS ImportLog."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-    )
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.status == ImportLog.Status.SUCCESS
-    assert log.records_imported == 1
-    assert log.records_failed == 0
-
-
-@pytest.mark.django_db
-def test_sync_crops_partial_success_import_log(monkeypatch) -> None:
-    """One crop saved + one failed produces a PARTIAL_SUCCESS ImportLog."""
-    # Two items; first succeeds, second has bad external_id (empty -> error)
-    mock_qs = Mock()
-    mock_qs.values.return_value = [
-        {"pk": 1, "external_id": "/api/v1/media/items/1"},
-        {"pk": 2, "external_id": ""},  # will be skipped as error
-    ]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}]}, None),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"data")
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "crops/img.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.return_value = (Mock(), True)
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.status == ImportLog.Status.PARTIAL_SUCCESS
-    assert log.records_imported == 1
-    assert log.records_failed == 1
-
-
-@pytest.mark.django_db
-def test_sync_crops_all_failed_import_log(monkeypatch) -> None:
-    """All items failing produces a FAILED ImportLog."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-        download_result=None,  # download always fails
-    )
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.status == ImportLog.Status.FAILED
-    assert log.records_imported == 0
-    assert log.records_failed == 1
-
-
-@pytest.mark.django_db
-def test_sync_crops_dry_run_does_not_write(monkeypatch) -> None:
-    """dry_run=True logs intentions but writes neither images nor DB rows."""
-    mock_crop_cls = _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-    )
-
-    result = sync_media_item_crops(dry_run=True)
-
-    assert result == 1  # "would save"
-    mock_crop_cls.objects.update_or_create.assert_not_called()
-    mock_crop_cls._meta.get_field.return_value.storage.save.assert_not_called()
-
-
-@pytest.mark.django_db
-def test_sync_crops_dry_run_logs_intended_save(monkeypatch, caplog) -> None:
-    """dry_run mode logs what would be saved for each crop."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-    )
-
-    caplog.set_level(logging.INFO, logger=viernulvier.logger.name)
-    sync_media_item_crops(dry_run=True)
-
-    assert any("[DRY RUN]" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_on_progress_called_per_item(monkeypatch) -> None:
-    """on_progress callback receives (idx, total) for every processed item."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[
-            {"pk": 1, "external_id": "/api/v1/media/items/1"},
-            {"pk": 2, "external_id": "/api/v1/media/items/2"},
-        ],
-        fetch_results={
-            "https://www.viernulvier.gent/api/v1/media/items/1": (
-                {"crops": [{"name": "hd_ready", "url": "https://cdn.example.com/a.jpg"}]},
-                None,
-            ),
-            "https://www.viernulvier.gent/api/v1/media/items/2": (
-                {"crops": [{"name": "hd_ready", "url": "https://cdn.example.com/b.jpg"}]},
-                None,
-            ),
-        },
-    )
-
-    calls = []
-    sync_media_item_crops(on_progress=lambda idx, total: calls.append((idx, total)))
-
-    assert len(calls) == 2
-    assert calls[0] == (1, 2)
-    assert calls[1] == (2, 2)
-
-
-@pytest.mark.django_db
-def test_sync_crops_save_exception_counted_as_error(monkeypatch, caplog) -> None:
-    """An unexpected exception during update_or_create is caught and counted as error."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}]}, None),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"data")
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "crops/img.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.side_effect = RuntimeError("db crash")
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    caplog.set_level(logging.ERROR, logger=viernulvier.logger.name)
-    result = sync_media_item_crops()
-
-    assert result == 0
-    assert any("Error saving crop" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_save_exception_added_to_error_messages(monkeypatch) -> None:
-    """Exceptions during save are captured in the ImportLog error_message."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}]}, None),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"data")
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "crops/img.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.side_effect = RuntimeError("db crash")
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.error_message is not None
-    assert "db crash" in log.error_message
-
-
-@pytest.mark.django_db
-def test_sync_crops_external_id_as_absolute_url_used_directly(monkeypatch) -> None:
-    """external_id already starting with http is used as-is (not prefixed)."""
-    captured_urls = []
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "https://www.viernulvier.gent/api/v1/media/items/99"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-
-    def fake_fetch(session, url, **_kw):
-        captured_urls.append(url)
-        return ({"crops": []}, None)
-
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", fake_fetch)
-
-    sync_media_item_crops()
-
-    assert captured_urls[0] == "https://www.viernulvier.gent/api/v1/media/items/99"
-
-
-@pytest.mark.django_db
-def test_sync_crops_external_id_as_relative_path_prefixed_with_base_domain(monkeypatch) -> None:
-    """Relative external_id is prefixed with BASE_DOMAIN before fetching."""
-    captured_urls = []
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/10"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-
-    def fake_fetch(session, url, **_kw):
-        captured_urls.append(url)
-        return ({"crops": []}, None)
-
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", fake_fetch)
-
-    sync_media_item_crops()
-
-    assert captured_urls[0].startswith("https://")
-    assert "/api/v1/media/items/10" in captured_urls[0]
-
-
-@pytest.mark.django_db
-def test_sync_crops_multiple_items_and_crops_counted_correctly(monkeypatch) -> None:
-    """Two items, each with two wanted crops -> four total crops saved."""
-    items_data = [
-        {"pk": 1, "external_id": "/api/v1/media/items/1"},
-        {"pk": 2, "external_id": "/api/v1/media/items/2"},
-    ]
-    mock_qs = Mock()
-    mock_qs.values.return_value = items_data
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (
-            {
-                "crops": [
-                    {"name": "hd_ready", "url": "https://cdn.example.com/hd.jpg"},
-                    {"name": "FE3_header", "url": "https://cdn.example.com/fe3.jpg"},
-                ]
-            },
-            None,
-        ),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"img")
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready", "FE3_header"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "crops/img.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.return_value = (Mock(), True)
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    result = sync_media_item_crops()
-
-    assert result == 4  # 2 items x 2 crops each
-
-
-@pytest.mark.django_db
-def test_sync_crops_crop_dict_not_dict_skipped(monkeypatch) -> None:
-    """Non-dict entries inside the crops list are silently ignored."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (
-            # Mix of valid and non-dict crops
-            {
-                "crops": [
-                    "not-a-dict",
-                    42,
-                    None,
-                    {"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"},
-                ]
-            },
-            None,
-        ),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"img")
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "crops/img.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.return_value = (Mock(), True)
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    result = sync_media_item_crops()
-
-    assert result == 1  # Only the valid dict crop was processed
-
-
-@pytest.mark.django_db
-def test_sync_crops_logs_created_or_updated_debug(monkeypatch, caplog) -> None:
-    """Created vs Updated status is logged at DEBUG level."""
-    mock_crop_cls = _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-    )
-    mock_crop_cls.objects.update_or_create.return_value = (Mock(), False)  # updated
-
-    caplog.set_level(logging.DEBUG, logger=viernulvier.logger.name)
-    sync_media_item_crops()
-
-    assert any("Updated" in r.message or "Created" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_import_log_source_is_correct(monkeypatch) -> None:
-    """The ImportLog source is set to 'viernulvier:media_item_crops'."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[],
-    )
-    mock_qs = Mock()
-    mock_qs.values.return_value = []
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.source == "viernulvier:media_item_crops"
-
-
-@pytest.mark.django_db
-def test_sync_crops_import_log_timestamps_sequential(monkeypatch) -> None:
-    """finished_at >= started_at in the ImportLog."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-    )
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert log.finished_at >= log.started_at
-
-
-@pytest.mark.django_db
-def test_sync_crops_fetch_error_message_captured_in_import_log(monkeypatch) -> None:
-    """ScraperError messages during fetch are included in the ImportLog error_message."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (_ for _ in ()).throw(ScraperError("network timeout")),
-    )
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert "network timeout" in (log.error_message or "")
-
-
-@pytest.mark.django_db
-def test_sync_crops_summary_logged_on_completion(monkeypatch, caplog) -> None:
-    """'Crop sync complete' with saved= and errors= is logged after running."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-    )
-
-    caplog.set_level(logging.INFO, logger=viernulvier.logger.name)
-    sync_media_item_crops()
-
-    assert any("Crop sync complete" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_dry_run_summary_has_dry_run_suffix(monkeypatch, caplog) -> None:
-    """The completion log includes '[DRY RUN]' suffix when dry_run=True."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-    )
-
-    caplog.set_level(logging.INFO, logger=viernulvier.logger.name)
-    sync_media_item_crops(dry_run=True)
-
-    assert any("[DRY RUN]" in r.message and "Crop sync complete" in r.message for r in caplog.records)
-
-
-@pytest.mark.django_db
-def test_sync_crops_download_error_message_captured(monkeypatch) -> None:
-    """Download failures are captured in the ImportLog error_message."""
-    _patch_crop_dependencies(
-        monkeypatch,
-        foto_items=[{"pk": 1, "external_id": "/api/v1/media/items/1"}],
-        download_result=None,
-    )
-
-    sync_media_item_crops()
-
-    log = ImportLog.objects.latest("started_at")
-    assert "Download failed" in (log.error_message or "")
-
-
-@pytest.mark.django_db
-def test_sync_crops_storage_save_called_with_image_bytes(monkeypatch) -> None:
-    """The storage backend's save() is called with a ContentFile wrapping the downloaded bytes."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}]}, None),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"raw-pixels")
-
-    storage_save_calls = []
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_storage = Mock()
-
-    def capture_save(upload_name, content_file):
-        storage_save_calls.append((upload_name, content_file.read()))
-        return upload_name
-
-    mock_storage.save.side_effect = capture_save
-    mock_image_field.storage = mock_storage
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-    mock_crop_cls.objects.update_or_create.return_value = (Mock(), True)
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    sync_media_item_crops()
-
-    assert storage_save_calls
-    assert storage_save_calls[0][1] == b"raw-pixels"
-
-
-# ============================================================================
-# on_progress called in every early-exit branch
-# ============================================================================
-
-
-@pytest.mark.django_db
-def test_on_progress_called_when_external_id_missing(monkeypatch) -> None:
-    """on_progress is invoked even when an item is skipped due to missing external_id."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": ""}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-
-    calls = []
-    sync_media_item_crops(on_progress=lambda idx, total: calls.append((idx, total)))
-
-    assert calls == [(1, 1)]
-
-
-@pytest.mark.django_db
-def test_on_progress_called_after_fetch_scraper_error(monkeypatch) -> None:
-    """on_progress is invoked after a ScraperError on the individual item fetch."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: (_ for _ in ()).throw(ScraperError("boom")),
-    )
-
-    calls = []
-    sync_media_item_crops(on_progress=lambda idx, total: calls.append((idx, total)))
-
-    assert calls == [(1, 1)]
-
-
-@pytest.mark.django_db
-def test_on_progress_called_after_304_not_modified(monkeypatch) -> None:
-    """on_progress is invoked when the item fetch returns None (304 Not Modified)."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(viernulvier, "_fetch_with_retry", lambda *_a, **_kw: (None, None))
-
-    calls = []
-    sync_media_item_crops(on_progress=lambda idx, total: calls.append((idx, total)))
-
-    assert calls == [(1, 1)]
-
-
-@pytest.mark.django_db
-def test_on_progress_called_after_non_list_crops(monkeypatch) -> None:
-    """on_progress is invoked when the crops field is not a list."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 1, "external_id": "/api/v1/media/items/1"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": "not-a-list"}, None),
-    )
-
-    calls = []
-    sync_media_item_crops(on_progress=lambda idx, total: calls.append((idx, total)))
-
-    assert calls == [(1, 1)]
-
-
-@pytest.mark.django_db
-def test_on_progress_covers_all_four_early_exit_branches_in_one_run(monkeypatch) -> None:
-    """Four items, each triggering a different early-exit; on_progress fired for all four."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [
-        {"pk": 1, "external_id": ""},  # branch: no external_id
-        {"pk": 2, "external_id": "/api/v1/media/items/2"},  # branch: ScraperError
-        {"pk": 3, "external_id": "/api/v1/media/items/3"},  # branch: 304 (None)
-        {"pk": 4, "external_id": "/api/v1/media/items/4"},  # branch: non-list crops
-    ]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-
-    calls = []
-    sync_media_item_crops(on_progress=lambda idx, total: calls.append((idx, total)))
-
-    assert calls == [(1, 4), (2, 4), (3, 4), (4, 4)]
-
-
-@pytest.mark.django_db
-def test_sync_crops_update_or_create_receives_correct_kwargs(monkeypatch) -> None:
-    """update_or_create is called with media_item_id, name, and defaults={image:...}."""
-    mock_qs = Mock()
-    mock_qs.values.return_value = [{"pk": 42, "external_id": "/api/v1/media/items/42"}]
-    monkeypatch.setattr(media_models.MediaItem.objects, "filter", lambda **_: mock_qs)
-    monkeypatch.setattr(viernulvier, "_build_session", Mock)
-    monkeypatch.setattr(
-        viernulvier,
-        "_fetch_with_retry",
-        lambda *_a, **_kw: ({"crops": [{"name": "hd_ready", "url": "https://cdn.example.com/img.jpg"}]}, None),
-    )
-    monkeypatch.setattr(viernulvier, "_download_image", lambda *_a: b"data")
-
-    uoc_calls = []
-
-    mock_crop_cls = Mock()
-    mock_crop_cls.SYNCED_CROP_NAMES = {"hd_ready"}
-    mock_image_field = Mock()
-    mock_image_field.generate_filename.return_value = "crops/img.jpg"
-    mock_image_field.storage = Mock()
-    mock_image_field.storage.save.return_value = "saved/path.jpg"
-    mock_crop_cls._meta = Mock()
-    mock_crop_cls._meta.get_field.return_value = mock_image_field
-
-    def capture_uoc(**kwargs):
-        uoc_calls.append(kwargs)
-        return (Mock(), True)
-
-    mock_crop_cls.objects.update_or_create.side_effect = capture_uoc
-    monkeypatch.setattr(media_models, "MediaItemCrop", mock_crop_cls)
-
-    sync_media_item_crops()
-
-    assert uoc_calls
-    call_kw = uoc_calls[0]
-    assert call_kw["media_item_id"] == 42
-    assert call_kw["name"] == "hd_ready"
-    assert call_kw["defaults"] == {"image": "saved/path.jpg"}
